LayoutTests/fast/canvas/rgba-parsing-expected.txt

Test rgba() color parsing results

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS parse('rgba(0, 0, 0, -0.10)') is 'RGBA[0, 0, 0, 0]'
PASS parse('rgba(0, 0, 0, -5.0)') is 'RGBA[0, 0, 0, 0]'
PASS parse('rgba(0, 0, 0, 5.0)') is 'RGBA[0, 0, 0, 255]'
PASS parse('rgba(0, 0, 0, -1)') is 'RGBA[0, 0, 0, 0]'
PASS parse('rgba(0, 0, 0, 0)') is 'RGBA[0, 0, 0, 0]'
PASS parse('rgba(0, 0, 0, 2)') is 'RGBA[0, 0, 0, 255]'
PASS parse('rgba(0, 0, 0, 2.0)') is 'RGBA[0, 0, 0, 255]'
PASS parse('rgba(0, 0, 0, 0.0)') is 'RGBA[0, 0, 0, 0]'
PASS parse('rgba(0, 0, 0, 00.0)') is 'RGBA[0, 0, 0, 0]'
PASS parse('rgba(0, 0, 0, 0.00)') is 'RGBA[0, 0, 0, 0]'
PASS parse('rgba(0, 0, 0, .1)') is 'RGBA[0, 0, 0, 25]'
PASS parse('rgba(0, 0, 0, .2)') is 'RGBA[0, 0, 0, 51]'
PASS parse('rgba(0, 0, 0, .3)') is 'RGBA[0, 0, 0, 76]'
PASS parse('rgba(0, 0, 0, .4)') is 'RGBA[0, 0, 0, 102]'
PASS parse('rgba(0, 0, 0, .5)') is 'RGBA[0, 0, 0, 127]'
PASS parse('rgba(0, 0, 0, .6)') is 'RGBA[0, 0, 0, 153]'
PASS parse('rgba(0, 0, 0, .7)') is 'RGBA[0, 0, 0, 179]'
PASS parse('rgba(0, 0, 0, .8)') is 'RGBA[0, 0, 0, 204]'
PASS parse('rgba(0, 0, 0, .9)') is 'RGBA[0, 0, 0, 230]'
PASS parse('rgba(0, 0, 0, 0.1)') is 'RGBA[0, 0, 0, 25]'
PASS parse('rgba(0, 0, 0, 0.2)') is 'RGBA[0, 0, 0, 51]'
PASS parse('rgba(0, 0, 0, 0.3)') is 'RGBA[0, 0, 0, 76]'
PASS parse('rgba(0, 0, 0, 0.4)') is 'RGBA[0, 0, 0, 102]'
PASS parse('rgba(0, 0, 0, 0.5)') is 'RGBA[0, 0, 0, 127]'
PASS parse('rgba(0, 0, 0, 0.6)') is 'RGBA[0, 0, 0, 153]'
PASS parse('rgba(0, 0, 0, 0.7)') is 'RGBA[0, 0, 0, 179]'
PASS parse('rgba(0, 0, 0, 0.8)') is 'RGBA[0, 0, 0, 204]'
PASS parse('rgba(0, 0, 0, 0.9)') is 'RGBA[0, 0, 0, 230]'
PASS parse('rgba(0, 0, 0, 1.0)') is 'RGBA[0, 0, 0, 255]'
PASS parse('rgba(0, 0, 0, 0.10)') is 'RGBA[0, 0, 0, 25]'
PASS parse('rgba(0, 0, 0, 0.20)') is 'RGBA[0, 0, 0, 51]'
PASS parse('rgba(0, 0, 0, 0.30)') is 'RGBA[0, 0, 0, 76]'
PASS parse('rgba(0, 0, 0, 0.40)') is 'RGBA[0, 0, 0, 102]'
PASS parse('rgba(0, 0, 0, 0.50)') is 'RGBA[0, 0, 0, 127]'
PASS parse('rgba(0, 0, 0, 0.60)') is 'RGBA[0, 0, 0, 153]'
PASS parse('rgba(0, 0, 0, 0.70)') is 'RGBA[0, 0, 0, 179]'
PASS parse('rgba(0, 0, 0, 0.80)') is 'RGBA[0, 0, 0, 204]'
PASS parse('rgba(0, 0, 0, 0.90)') is 'RGBA[0, 0, 0, 230]'
PASS parse('rgba(0, 0, 0, 1.00)') is 'RGBA[0, 0, 0, 255]'
PASS parse('rgba(0, 0, 0, .10)') is 'RGBA[0, 0, 0, 25]'
PASS parse('rgba(0, 0, 0, .20)') is 'RGBA[0, 0, 0, 51]'
PASS parse('rgba(0, 0, 0, .30)') is 'RGBA[0, 0, 0, 76]'
PASS parse('rgba(0, 0, 0, .40)') is 'RGBA[0, 0, 0, 102]'
PASS parse('rgba(0, 0, 0, .50)') is 'RGBA[0, 0, 0, 127]'
PASS parse('rgba(0, 0, 0, .60)') is 'RGBA[0, 0, 0, 153]'
PASS parse('rgba(0, 0, 0, .70)') is 'RGBA[0, 0, 0, 179]'
PASS parse('rgba(0, 0, 0, .80)') is 'RGBA[0, 0, 0, 204]'
PASS parse('rgba(0, 0, 0, .90)') is 'RGBA[0, 0, 0, 230]'
PASS parse('rgba(0, 0, 0, 0.10000000000000000000000)') is 'RGBA[0, 0, 0, 25]'
PASS parse('rgba(0, 0, 0, 0.20000000000000000000000)') is 'RGBA[0, 0, 0, 51]'
PASS parse('rgba(0, 0, 0, 0.30000000000000000000000)') is 'RGBA[0, 0, 0, 76]'
PASS parse('rgba(0, 0, 0, 0.40000000000000000000000)') is 'RGBA[0, 0, 0, 102]'
PASS parse('rgba(0, 0, 0, 0.50000000000000000000000)') is 'RGBA[0, 0, 0, 127]'
PASS parse('rgba(0, 0, 0, 0.60000000000000000000000)') is 'RGBA[0, 0, 0, 153]'
PASS parse('rgba(0, 0, 0, 0.70000000000000000000000)') is 'RGBA[0, 0, 0, 179]'
PASS parse('rgba(0, 0, 0, 0.80000000000000000000000)') is 'RGBA[0, 0, 0, 204]'
PASS parse('rgba(0, 0, 0, 0.90000000000000000000000)') is 'RGBA[0, 0, 0, 230]'
PASS parse('rgba(0, 0, 0, 1.00000000000000000000000)') is 'RGBA[0, 0, 0, 255]'
PASS parse('rgba(0, 0, 0, 0.990)') is 'RGBA[0, 0, 0, 253]'
PASS parse('rgba(0, 0, 0, 0.991)') is 'RGBA[0, 0, 0, 253]'
PASS parse('rgba(0, 0, 0, 0.992)') is 'RGBA[0, 0, 0, 253]'
PASS parse('rgba(0, 0, 0, 0.993)') is 'RGBA[0, 0, 0, 254]'
PASS parse('rgba(0, 0, 0, 0.994)') is 'RGBA[0, 0, 0, 254]'
PASS parse('rgba(0, 0, 0, 0.995)') is 'RGBA[0, 0, 0, 254]'
PASS parse('rgba(0, 0, 0, 0.996)') is 'RGBA[0, 0, 0, 254]'
PASS parse('rgba(0, 0, 0, 0.997)') is 'RGBA[0, 0, 0, 255]'
PASS parse('rgba(0, 0, 0, 0.998)') is 'RGBA[0, 0, 0, 255]'
PASS parse('rgba(0, 0, 0, 0.999)') is 'RGBA[0, 0, 0, 255]'
rgba(0, 0, 0, 0) => RGBA[0, 0, 0, 0]
rgba(0, 0, 0, 0.01333) => RGBA[0, 0, 0, 3]
rgba(0, 0, 0, 0.02666) => RGBA[0, 0, 0, 6]
rgba(0, 0, 0, 0.03999) => RGBA[0, 0, 0, 10]
rgba(0, 0, 0, 0.05332) => RGBA[0, 0, 0, 13]
rgba(0, 0, 0, 0.06665) => RGBA[0, 0, 0, 17]
rgba(0, 0, 0, 0.07998) => RGBA[0, 0, 0, 20]
rgba(0, 0, 0, 0.09330999999999999) => RGBA[0, 0, 0, 23]
rgba(0, 0, 0, 0.10663999999999998) => RGBA[0, 0, 0, 27]
rgba(0, 0, 0, 0.11996999999999998) => RGBA[0, 0, 0, 30]
rgba(0, 0, 0, 0.13329999999999997) => RGBA[0, 0, 0, 34]
rgba(0, 0, 0, 0.14662999999999998) => RGBA[0, 0, 0, 37]
rgba(0, 0, 0, 0.15996) => RGBA[0, 0, 0, 40]
rgba(0, 0, 0, 0.17329) => RGBA[0, 0, 0, 44]
rgba(0, 0, 0, 0.18662) => RGBA[0, 0, 0, 47]
rgba(0, 0, 0, 0.19995000000000002) => RGBA[0, 0, 0, 51]
rgba(0, 0, 0, 0.21328000000000003) => RGBA[0, 0, 0, 54]
rgba(0, 0, 0, 0.22661000000000003) => RGBA[0, 0, 0, 58]
rgba(0, 0, 0, 0.23994000000000004) => RGBA[0, 0, 0, 61]
rgba(0, 0, 0, 0.25327000000000005) => RGBA[0, 0, 0, 64]
rgba(0, 0, 0, 0.26660000000000006) => RGBA[0, 0, 0, 68]
rgba(0, 0, 0, 0.27993000000000007) => RGBA[0, 0, 0, 71]
rgba(0, 0, 0, 0.2932600000000001) => RGBA[0, 0, 0, 75]
rgba(0, 0, 0, 0.3065900000000001) => RGBA[0, 0, 0, 78]
rgba(0, 0, 0, 0.3199200000000001) => RGBA[0, 0, 0, 81]
rgba(0, 0, 0, 0.3332500000000001) => RGBA[0, 0, 0, 85]
rgba(0, 0, 0, 0.3465800000000001) => RGBA[0, 0, 0, 88]
rgba(0, 0, 0, 0.3599100000000001) => RGBA[0, 0, 0, 92]
rgba(0, 0, 0, 0.3732400000000001) => RGBA[0, 0, 0, 95]
rgba(0, 0, 0, 0.38657000000000014) => RGBA[0, 0, 0, 98]
rgba(0, 0, 0, 0.39990000000000014) => RGBA[0, 0, 0, 102]
rgba(0, 0, 0, 0.41323000000000015) => RGBA[0, 0, 0, 105]
rgba(0, 0, 0, 0.42656000000000016) => RGBA[0, 0, 0, 109]
rgba(0, 0, 0, 0.43989000000000017) => RGBA[0, 0, 0, 112]
rgba(0, 0, 0, 0.4532200000000002) => RGBA[0, 0, 0, 116]
rgba(0, 0, 0, 0.4665500000000002) => RGBA[0, 0, 0, 119]
rgba(0, 0, 0, 0.4798800000000002) => RGBA[0, 0, 0, 122]
rgba(0, 0, 0, 0.4932100000000002) => RGBA[0, 0, 0, 126]
rgba(0, 0, 0, 0.5065400000000002) => RGBA[0, 0, 0, 129]
rgba(0, 0, 0, 0.5198700000000002) => RGBA[0, 0, 0, 133]
rgba(0, 0, 0, 0.5332000000000001) => RGBA[0, 0, 0, 136]
rgba(0, 0, 0, 0.5465300000000001) => RGBA[0, 0, 0, 139]
rgba(0, 0, 0, 0.55986) => RGBA[0, 0, 0, 143]
rgba(0, 0, 0, 0.57319) => RGBA[0, 0, 0, 146]
rgba(0, 0, 0, 0.5865199999999999) => RGBA[0, 0, 0, 150]
rgba(0, 0, 0, 0.5998499999999999) => RGBA[0, 0, 0, 153]
rgba(0, 0, 0, 0.6131799999999998) => RGBA[0, 0, 0, 156]
rgba(0, 0, 0, 0.6265099999999998) => RGBA[0, 0, 0, 160]
rgba(0, 0, 0, 0.6398399999999997) => RGBA[0, 0, 0, 163]
rgba(0, 0, 0, 0.6531699999999997) => RGBA[0, 0, 0, 167]
rgba(0, 0, 0, 0.6664999999999996) => RGBA[0, 0, 0, 170]
rgba(0, 0, 0, 0.6798299999999996) => RGBA[0, 0, 0, 174]
rgba(0, 0, 0, 0.6931599999999996) => RGBA[0, 0, 0, 177]
rgba(0, 0, 0, 0.7064899999999995) => RGBA[0, 0, 0, 180]
rgba(0, 0, 0, 0.7198199999999995) => RGBA[0, 0, 0, 184]
rgba(0, 0, 0, 0.7331499999999994) => RGBA[0, 0, 0, 187]
rgba(0, 0, 0, 0.7464799999999994) => RGBA[0, 0, 0, 191]
rgba(0, 0, 0, 0.7598099999999993) => RGBA[0, 0, 0, 194]
rgba(0, 0, 0, 0.7731399999999993) => RGBA[0, 0, 0, 197]
rgba(0, 0, 0, 0.7864699999999992) => RGBA[0, 0, 0, 201]
rgba(0, 0, 0, 0.7997999999999992) => RGBA[0, 0, 0, 204]
rgba(0, 0, 0, 0.8131299999999991) => RGBA[0, 0, 0, 208]
rgba(0, 0, 0, 0.8264599999999991) => RGBA[0, 0, 0, 211]
rgba(0, 0, 0, 0.839789999999999) => RGBA[0, 0, 0, 214]
rgba(0, 0, 0, 0.853119999999999) => RGBA[0, 0, 0, 218]
rgba(0, 0, 0, 0.8664499999999989) => RGBA[0, 0, 0, 221]
rgba(0, 0, 0, 0.8797799999999989) => RGBA[0, 0, 0, 225]
rgba(0, 0, 0, 0.8931099999999988) => RGBA[0, 0, 0, 228]
rgba(0, 0, 0, 0.9064399999999988) => RGBA[0, 0, 0, 232]
rgba(0, 0, 0, 0.9197699999999988) => RGBA[0, 0, 0, 235]
rgba(0, 0, 0, 0.9330999999999987) => RGBA[0, 0, 0, 238]
rgba(0, 0, 0, 0.9464299999999987) => RGBA[0, 0, 0, 242]
rgba(0, 0, 0, 0.9597599999999986) => RGBA[0, 0, 0, 245]
rgba(0, 0, 0, 0.9730899999999986) => RGBA[0, 0, 0, 249]
rgba(0, 0, 0, 0.9864199999999985) => RGBA[0, 0, 0, 252]
rgba(0, 0, 0, 0.9997499999999985) => RGBA[0, 0, 0, 255]
rgba(63, 63, 63, 0) => RGBA[0, 0, 0, 0]
rgba(63, 63, 63, 0.01333) => RGBA[85, 85, 85, 3]
rgba(63, 63, 63, 0.02666) => RGBA[42, 42, 42, 6]
rgba(63, 63, 63, 0.03999) => RGBA[51, 51, 51, 10]
rgba(63, 63, 63, 0.05332) => RGBA[58, 58, 58, 13]
rgba(63, 63, 63, 0.06665) => RGBA[60, 60, 60, 17]
rgba(63, 63, 63, 0.07998) => RGBA[63, 63, 63, 20]
rgba(63, 63, 63, 0.09330999999999999) => RGBA[66, 66, 66, 23]
rgba(63, 63, 63, 0.10663999999999998) => RGBA[66, 66, 66, 27]
rgba(63, 63, 63, 0.11996999999999998) => RGBA[59, 59, 59, 30]
rgba(63, 63, 63, 0.13329999999999997) => RGBA[60, 60, 60, 34]
rgba(63, 63, 63, 0.14662999999999998) => RGBA[62, 62, 62, 37]
rgba(63, 63, 63, 0.15996) => RGBA[63, 63, 63, 40]
rgba(63, 63, 63, 0.17329) => RGBA[63, 63, 63, 44]
rgba(63, 63, 63, 0.18662) => RGBA[65, 65, 65, 47]
rgba(63, 63, 63, 0.19995000000000002) => RGBA[65, 65, 65, 51]
rgba(63, 63, 63, 0.21328000000000003) => RGBA[61, 61, 61, 54]
rgba(63, 63, 63, 0.22661000000000003) => RGBA[61, 61, 61, 58]
rgba(63, 63, 63, 0.23994000000000004) => RGBA[62, 62, 62, 61]
rgba(63, 63, 63, 0.25327000000000005) => RGBA[63, 63, 63, 64]
rgba(63, 63, 63, 0.26660000000000006) => RGBA[63, 63, 63, 68]
rgba(63, 63, 63, 0.27993000000000007) => RGBA[64, 64, 64, 71]
rgba(63, 63, 63, 0.2932600000000001) => RGBA[64, 64, 64, 75]
rgba(63, 63, 63, 0.3065900000000001) => RGBA[62, 62, 62, 78]
rgba(63, 63, 63, 0.3199200000000001) => RGBA[62, 62, 62, 81]
rgba(63, 63, 63, 0.3332500000000001) => RGBA[63, 63, 63, 85]
rgba(63, 63, 63, 0.3465800000000001) => RGBA[63, 63, 63, 88]
rgba(63, 63, 63, 0.3599100000000001) => RGBA[63, 63, 63, 92]
rgba(63, 63, 63, 0.3732400000000001) => RGBA[61, 61, 61, 95]
rgba(63, 63, 63, 0.38657000000000014) => RGBA[62, 62, 62, 98]
rgba(63, 63, 63, 0.39990000000000014) => RGBA[62, 62, 62, 102]
rgba(63, 63, 63, 0.41323000000000015) => RGBA[63, 63, 63, 105]
rgba(63, 63, 63, 0.42656000000000016) => RGBA[63, 63, 63, 109]
rgba(63, 63, 63, 0.43989000000000017) => RGBA[63, 63, 63, 112]
rgba(63, 63, 63, 0.4532200000000002) => RGBA[63, 63, 63, 116]
rgba(63, 63, 63, 0.4665500000000002) => RGBA[62, 62, 62, 119]
rgba(63, 63, 63, 0.4798800000000002) => RGBA[62, 62, 62, 122]
rgba(63, 63, 63, 0.4932100000000002) => RGBA[62, 62, 62, 126]
rgba(63, 63, 63, 0.5065400000000002) => RGBA[63, 63, 63, 129]
rgba(63, 63, 63, 0.5198700000000002) => RGBA[63, 63, 63, 133]
rgba(63, 63, 63, 0.5332000000000001) => RGBA[63, 63, 63, 136]
rgba(63, 63, 63, 0.5465300000000001) => RGBA[62, 62, 62, 139]
rgba(63, 63, 63, 0.55986) => RGBA[62, 62, 62, 143]
rgba(63, 63, 63, 0.57319) => RGBA[62, 62, 62, 146]
rgba(63, 63, 63, 0.5865199999999999) => RGBA[62, 62, 62, 150]
rgba(63, 63, 63, 0.5998499999999999) => RGBA[63, 63, 63, 153]
rgba(63, 63, 63, 0.6131799999999998) => RGBA[63, 63, 63, 156]
rgba(63, 63, 63, 0.6265099999999998) => RGBA[63, 63, 63, 160]
rgba(63, 63, 63, 0.6398399999999997) => RGBA[62, 62, 62, 163]
rgba(63, 63, 63, 0.6531699999999997) => RGBA[62, 62, 62, 167]
rgba(63, 63, 63, 0.6664999999999996) => RGBA[63, 63, 63, 170]
rgba(63, 63, 63, 0.6798299999999996) => RGBA[63, 63, 63, 174]
rgba(63, 63, 63, 0.6931599999999996) => RGBA[63, 63, 63, 177]
rgba(63, 63, 63, 0.7064899999999995) => RGBA[62, 62, 62, 180]
rgba(63, 63, 63, 0.7198199999999995) => RGBA[62, 62, 62, 184]
rgba(63, 63, 63, 0.7331499999999994) => RGBA[62, 62, 62, 187]
rgba(63, 63, 63, 0.7464799999999994) => RGBA[62, 62, 62, 191]
rgba(63, 63, 63, 0.7598099999999993) => RGBA[63, 63, 63, 194]
rgba(63, 63, 63, 0.7731399999999993) => RGBA[63, 63, 63, 197]
rgba(63, 63, 63, 0.7864699999999992) => RGBA[63, 63, 63, 201]
rgba(63, 63, 63, 0.7997999999999992) => RGBA[62, 62, 62, 204]
rgba(63, 63, 63, 0.8131299999999991) => RGBA[62, 62, 62, 208]
rgba(63, 63, 63, 0.8264599999999991) => RGBA[62, 62, 62, 211]
rgba(63, 63, 63, 0.839789999999999) => RGBA[63, 63, 63, 214]
rgba(63, 63, 63, 0.853119999999999) => RGBA[63, 63, 63, 218]
rgba(63, 63, 63, 0.8664499999999989) => RGBA[63, 63, 63, 221]
rgba(63, 63, 63, 0.8797799999999989) => RGBA[63, 63, 63, 225]
rgba(63, 63, 63, 0.8931099999999988) => RGBA[62, 62, 62, 228]
rgba(63, 63, 63, 0.9064399999999988) => RGBA[62, 62, 62, 232]
rgba(63, 63, 63, 0.9197699999999988) => RGBA[62, 62, 62, 235]
rgba(63, 63, 63, 0.9330999999999987) => RGBA[63, 63, 63, 238]
rgba(63, 63, 63, 0.9464299999999987) => RGBA[63, 63, 63, 242]
rgba(63, 63, 63, 0.9597599999999986) => RGBA[63, 63, 63, 245]
rgba(63, 63, 63, 0.9730899999999986) => RGBA[63, 63, 63, 249]
rgba(63, 63, 63, 0.9864199999999985) => RGBA[62, 62, 62, 252]
rgba(63, 63, 63, 0.9997499999999985) => RGBA[63, 63, 63, 255]
rgba(126, 126, 126, 0) => RGBA[0, 0, 0, 0]
rgba(126, 126, 126, 0.01333) => RGBA[85, 85, 85, 3]
rgba(126, 126, 126, 0.02666) => RGBA[127, 127, 127, 6]
rgba(126, 126, 126, 0.03999) => RGBA[127, 127, 127, 10]
rgba(126, 126, 126, 0.05332) => RGBA[117, 117, 117, 13]
rgba(126, 126, 126, 0.06665) => RGBA[120, 120, 120, 17]
rgba(126, 126, 126, 0.07998) => RGBA[127, 127, 127, 20]
rgba(126, 126, 126, 0.09330999999999999) => RGBA[121, 121, 121, 23]
rgba(126, 126, 126, 0.10663999999999998) => RGBA[122, 122, 122, 27]
rgba(126, 126, 126, 0.11996999999999998) => RGBA[127, 127, 127, 30]
rgba(126, 126, 126, 0.13329999999999997) => RGBA[127, 127, 127, 34]
rgba(126, 126, 126, 0.14662999999999998) => RGBA[124, 124, 124, 37]
rgba(126, 126, 126, 0.15996) => RGBA[127, 127, 127, 40]
rgba(126, 126, 126, 0.17329) => RGBA[127, 127, 127, 44]
rgba(126, 126, 126, 0.18662) => RGBA[124, 124, 124, 47]
rgba(126, 126, 126, 0.19995000000000002) => RGBA[125, 125, 125, 51]
rgba(126, 126, 126, 0.21328000000000003) => RGBA[127, 127, 127, 54]
rgba(126, 126, 126, 0.22661000000000003) => RGBA[127, 127, 127, 58]
rgba(126, 126, 126, 0.23994000000000004) => RGBA[125, 125, 125, 61]
rgba(126, 126, 126, 0.25327000000000005) => RGBA[127, 127, 127, 64]
rgba(126, 126, 126, 0.26660000000000006) => RGBA[127, 127, 127, 68]
rgba(126, 126, 126, 0.27993000000000007) => RGBA[125, 125, 125, 71]
rgba(126, 126, 126, 0.2932600000000001) => RGBA[125, 125, 125, 75]
rgba(126, 126, 126, 0.3065900000000001) => RGBA[127, 127, 127, 78]
rgba(126, 126, 126, 0.3199200000000001) => RGBA[125, 125, 125, 81]
rgba(126, 126, 126, 0.3332500000000001) => RGBA[126, 126, 126, 85]
rgba(126, 126, 126, 0.3465800000000001) => RGBA[124, 124, 124, 88]
rgba(126, 126, 126, 0.3599100000000001) => RGBA[124, 124, 124, 92]
rgba(126, 126, 126, 0.3732400000000001) => RGBA[126, 126, 126, 95]
rgba(126, 126, 126, 0.38657000000000014) => RGBA[124, 124, 124, 98]
rgba(126, 126, 126, 0.39990000000000014) => RGBA[125, 125, 125, 102]
rgba(126, 126, 126, 0.41323000000000015) => RGBA[126, 126, 126, 105]
rgba(126, 126, 126, 0.42656000000000016) => RGBA[126, 126, 126, 109]
rgba(126, 126, 126, 0.43989000000000017) => RGBA[125, 125, 125, 112]
rgba(126, 126, 126, 0.4532200000000002) => RGBA[125, 125, 125, 116]
rgba(126, 126, 126, 0.4665500000000002) => RGBA[126, 126, 126, 119]
rgba(126, 126, 126, 0.4798800000000002) => RGBA[125, 125, 125, 122]
rgba(126, 126, 126, 0.4932100000000002) => RGBA[125, 125, 125, 126]
rgba(126, 126, 126, 0.5065400000000002) => RGBA[126, 126, 126, 129]
rgba(126, 126, 126, 0.5198700000000002) => RGBA[126, 126, 126, 133]
rgba(126, 126, 126, 0.5332000000000001) => RGBA[125, 125, 125, 136]
rgba(126, 126, 126, 0.5465300000000001) => RGBA[126, 126, 126, 139]
rgba(126, 126, 126, 0.55986) => RGBA[126, 126, 126, 143]
rgba(126, 126, 126, 0.57319) => RGBA[125, 125, 125, 146]
rgba(126, 126, 126, 0.5865199999999999) => RGBA[125, 125, 125, 150]
rgba(126, 126, 126, 0.5998499999999999) => RGBA[126, 126, 126, 153]
rgba(126, 126, 126, 0.6131799999999998) => RGBA[125, 125, 125, 156]
rgba(126, 126, 126, 0.6265099999999998) => RGBA[125, 125, 125, 160]
rgba(126, 126, 126, 0.6398399999999997) => RGBA[126, 126, 126, 163]
rgba(126, 126, 126, 0.6531699999999997) => RGBA[126, 126, 126, 167]
rgba(126, 126, 126, 0.6664999999999996) => RGBA[126, 126, 126, 170]
rgba(126, 126, 126, 0.6798299999999996) => RGBA[126, 126, 126, 174]
rgba(126, 126, 126, 0.6931599999999996) => RGBA[125, 125, 125, 177]
rgba(126, 126, 126, 0.7064899999999995) => RGBA[126, 126, 126, 180]
rgba(126, 126, 126, 0.7198199999999995) => RGBA[126, 126, 126, 184]
rgba(126, 126, 126, 0.7331499999999994) => RGBA[125, 125, 125, 187]
rgba(126, 126, 126, 0.7464799999999994) => RGBA[125, 125, 125, 191]
rgba(126, 126, 126, 0.7598099999999993) => RGBA[126, 126, 126, 194]
rgba(126, 126, 126, 0.7731399999999993) => RGBA[125, 125, 125, 197]
rgba(126, 126, 126, 0.7864699999999992) => RGBA[125, 125, 125, 201]
rgba(126, 126, 126, 0.7997999999999992) => RGBA[126, 126, 126, 204]
rgba(126, 126, 126, 0.8131299999999991) => RGBA[126, 126, 126, 208]
rgba(126, 126, 126, 0.8264599999999991) => RGBA[125, 125, 125, 211]
rgba(126, 126, 126, 0.839789999999999) => RGBA[126, 126, 126, 214]
rgba(126, 126, 126, 0.853119999999999) => RGBA[126, 126, 126, 218]
rgba(126, 126, 126, 0.8664499999999989) => RGBA[125, 125, 125, 221]
rgba(126, 126, 126, 0.8797799999999989) => RGBA[125, 125, 125, 225]
rgba(126, 126, 126, 0.8931099999999988) => RGBA[126, 126, 126, 228]
rgba(126, 126, 126, 0.9064399999999988) => RGBA[126, 126, 126, 232]
rgba(126, 126, 126, 0.9197699999999988) => RGBA[125, 125, 125, 235]
rgba(126, 126, 126, 0.9330999999999987) => RGBA[126, 126, 126, 238]
rgba(126, 126, 126, 0.9464299999999987) => RGBA[126, 126, 126, 242]
rgba(126, 126, 126, 0.9597599999999986) => RGBA[125, 125, 125, 245]
rgba(126, 126, 126, 0.9730899999999986) => RGBA[125, 125, 125, 249]
rgba(126, 126, 126, 0.9864199999999985) => RGBA[126, 126, 126, 252]
rgba(126, 126, 126, 0.9997499999999985) => RGBA[126, 126, 126, 255]
rgba(189, 189, 189, 0) => RGBA[0, 0, 0, 0]
rgba(189, 189, 189, 0.01333) => RGBA[170, 170, 170, 3]
rgba(189, 189, 189, 0.02666) => RGBA[170, 170, 170, 6]
rgba(189, 189, 189, 0.03999) => RGBA[178, 178, 178, 10]
rgba(189, 189, 189, 0.05332) => RGBA[196, 196, 196, 13]
rgba(189, 189, 189, 0.06665) => RGBA[195, 195, 195, 17]
rgba(189, 189, 189, 0.07998) => RGBA[191, 191, 191, 20]
rgba(189, 189, 189, 0.09330999999999999) => RGBA[188, 188, 188, 23]
rgba(189, 189, 189, 0.10663999999999998) => RGBA[188, 188, 188, 27]
rgba(189, 189, 189, 0.11996999999999998) => RGBA[187, 187, 187, 30]
rgba(189, 189, 189, 0.13329999999999997) => RGBA[187, 187, 187, 34]
rgba(189, 189, 189, 0.14662999999999998) => RGBA[186, 186, 186, 37]
rgba(189, 189, 189, 0.15996) => RGBA[191, 191, 191, 40]
rgba(189, 189, 189, 0.17329) => RGBA[191, 191, 191, 44]
rgba(189, 189, 189, 0.18662) => RGBA[189, 189, 189, 47]
rgba(189, 189, 189, 0.19995000000000002) => RGBA[190, 190, 190, 51]
rgba(189, 189, 189, 0.21328000000000003) => RGBA[188, 188, 188, 54]
rgba(189, 189, 189, 0.22661000000000003) => RGBA[189, 189, 189, 58]
rgba(189, 189, 189, 0.23994000000000004) => RGBA[188, 188, 188, 61]
rgba(189, 189, 189, 0.25327000000000005) => RGBA[187, 187, 187, 64]
rgba(189, 189, 189, 0.26660000000000006) => RGBA[187, 187, 187, 68]
rgba(189, 189, 189, 0.27993000000000007) => RGBA[190, 190, 190, 71]
rgba(189, 189, 189, 0.2932600000000001) => RGBA[190, 190, 190, 75]
rgba(189, 189, 189, 0.3065900000000001) => RGBA[189, 189, 189, 78]
rgba(189, 189, 189, 0.3199200000000001) => RGBA[188, 188, 188, 81]
rgba(189, 189, 189, 0.3332500000000001) => RGBA[189, 189, 189, 85]
rgba(189, 189, 189, 0.3465800000000001) => RGBA[188, 188, 188, 88]
rgba(189, 189, 189, 0.3599100000000001) => RGBA[188, 188, 188, 92]
rgba(189, 189, 189, 0.3732400000000001) => RGBA[187, 187, 187, 95]
rgba(189, 189, 189, 0.38657000000000014) => RGBA[189, 189, 189, 98]
rgba(189, 189, 189, 0.39990000000000014) => RGBA[190, 190, 190, 102]
rgba(189, 189, 189, 0.41323000000000015) => RGBA[189, 189, 189, 105]
rgba(189, 189, 189, 0.42656000000000016) => RGBA[189, 189, 189, 109]
rgba(189, 189, 189, 0.43989000000000017) => RGBA[188, 188, 188, 112]
rgba(189, 189, 189, 0.4532200000000002) => RGBA[189, 189, 189, 116]
rgba(189, 189, 189, 0.4665500000000002) => RGBA[188, 188, 188, 119]
rgba(189, 189, 189, 0.4798800000000002) => RGBA[188, 188, 188, 122]
rgba(189, 189, 189, 0.4932100000000002) => RGBA[188, 188, 188, 126]
rgba(189, 189, 189, 0.5065400000000002) => RGBA[189, 189, 189, 129]
rgba(189, 189, 189, 0.5198700000000002) => RGBA[189, 189, 189, 133]
rgba(189, 189, 189, 0.5332000000000001) => RGBA[189, 189, 189, 136]
rgba(189, 189, 189, 0.5465300000000001) => RGBA[188, 188, 188, 139]
rgba(189, 189, 189, 0.55986) => RGBA[189, 189, 189, 143]
rgba(189, 189, 189, 0.57319) => RGBA[188, 188, 188, 146]
rgba(189, 189, 189, 0.5865199999999999) => RGBA[188, 188, 188, 150]
rgba(189, 189, 189, 0.5998499999999999) => RGBA[188, 188, 188, 153]
rgba(189, 189, 189, 0.6131799999999998) => RGBA[189, 189, 189, 156]
rgba(189, 189, 189, 0.6265099999999998) => RGBA[189, 189, 189, 160]
rgba(189, 189, 189, 0.6398399999999997) => RGBA[189, 189, 189, 163]
rgba(189, 189, 189, 0.6531699999999997) => RGBA[189, 189, 189, 167]
rgba(189, 189, 189, 0.6664999999999996) => RGBA[189, 189, 189, 170]
rgba(189, 189, 189, 0.6798299999999996) => RGBA[189, 189, 189, 174]
rgba(189, 189, 189, 0.6931599999999996) => RGBA[188, 188, 188, 177]
rgba(189, 189, 189, 0.7064899999999995) => RGBA[188, 188, 188, 180]
rgba(189, 189, 189, 0.7198199999999995) => RGBA[188, 188, 188, 184]
rgba(189, 189, 189, 0.7331499999999994) => RGBA[189, 189, 189, 187]
rgba(189, 189, 189, 0.7464799999999994) => RGBA[189, 189, 189, 191]
rgba(189, 189, 189, 0.7598099999999993) => RGBA[189, 189, 189, 194]
rgba(189, 189, 189, 0.7731399999999993) => RGBA[188, 188, 188, 197]
rgba(189, 189, 189, 0.7864699999999992) => RGBA[189, 189, 189, 201]
rgba(189, 189, 189, 0.7997999999999992) => RGBA[188, 188, 188, 204]
rgba(189, 189, 189, 0.8131299999999991) => RGBA[188, 188, 188, 208]
rgba(189, 189, 189, 0.8264599999999991) => RGBA[188, 188, 188, 211]
rgba(189, 189, 189, 0.839789999999999) => RGBA[189, 189, 189, 214]
rgba(189, 189, 189, 0.853119999999999) => RGBA[189, 189, 189, 218]
rgba(189, 189, 189, 0.8664499999999989) => RGBA[189, 189, 189, 221]
rgba(189, 189, 189, 0.8797799999999989) => RGBA[189, 189, 189, 225]
rgba(189, 189, 189, 0.8931099999999988) => RGBA[189, 189, 189, 228]
rgba(189, 189, 189, 0.9064399999999988) => RGBA[189, 189, 189, 232]
rgba(189, 189, 189, 0.9197699999999988) => RGBA[188, 188, 188, 235]
rgba(189, 189, 189, 0.9330999999999987) => RGBA[188, 188, 188, 238]
rgba(189, 189, 189, 0.9464299999999987) => RGBA[188, 188, 188, 242]
rgba(189, 189, 189, 0.9597599999999986) => RGBA[189, 189, 189, 245]
rgba(189, 189, 189, 0.9730899999999986) => RGBA[189, 189, 189, 249]
rgba(189, 189, 189, 0.9864199999999985) => RGBA[189, 189, 189, 252]
rgba(189, 189, 189, 0.9997499999999985) => RGBA[189, 189, 189, 255]
rgba(252, 252, 252, 0) => RGBA[0, 0, 0, 0]
rgba(252, 252, 252, 0.01333) => RGBA[255, 255, 255, 3]
rgba(252, 252, 252, 0.02666) => RGBA[255, 255, 255, 6]
rgba(252, 252, 252, 0.03999) => RGBA[255, 255, 255, 10]
rgba(252, 252, 252, 0.05332) => RGBA[255, 255, 255, 13]
rgba(252, 252, 252, 0.06665) => RGBA[255, 255, 255, 17]
rgba(252, 252, 252, 0.07998) => RGBA[255, 255, 255, 20]
rgba(252, 252, 252, 0.09330999999999999) => RGBA[255, 255, 255, 23]
rgba(252, 252, 252, 0.10663999999999998) => RGBA[255, 255, 255, 27]
rgba(252, 252, 252, 0.11996999999999998) => RGBA[255, 255, 255, 30]
rgba(252, 252, 252, 0.13329999999999997) => RGBA[255, 255, 255, 34]
rgba(252, 252, 252, 0.14662999999999998) => RGBA[255, 255, 255, 37]
rgba(252, 252, 252, 0.15996) => RGBA[255, 255, 255, 40]
rgba(252, 252, 252, 0.17329) => RGBA[249, 249, 249, 44]
rgba(252, 252, 252, 0.18662) => RGBA[249, 249, 249, 47]
rgba(252, 252, 252, 0.19995000000000002) => RGBA[250, 250, 250, 51]
rgba(252, 252, 252, 0.21328000000000003) => RGBA[250, 250, 250, 54]
rgba(252, 252, 252, 0.22661000000000003) => RGBA[250, 250, 250, 58]
rgba(252, 252, 252, 0.23994000000000004) => RGBA[250, 250, 250, 61]
rgba(252, 252, 252, 0.25327000000000005) => RGBA[251, 251, 251, 64]
rgba(252, 252, 252, 0.26660000000000006) => RGBA[251, 251, 251, 68]
rgba(252, 252, 252, 0.27993000000000007) => RGBA[251, 251, 251, 71]
rgba(252, 252, 252, 0.2932600000000001) => RGBA[251, 251, 251, 75]
rgba(252, 252, 252, 0.3065900000000001) => RGBA[251, 251, 251, 78]
rgba(252, 252, 252, 0.3199200000000001) => RGBA[251, 251, 251, 81]
rgba(252, 252, 252, 0.3332500000000001) => RGBA[252, 252, 252, 85]
rgba(252, 252, 252, 0.3465800000000001) => RGBA[252, 252, 252, 88]
rgba(252, 252, 252, 0.3599100000000001) => RGBA[252, 252, 252, 92]
rgba(252, 252, 252, 0.3732400000000001) => RGBA[252, 252, 252, 95]
rgba(252, 252, 252, 0.38657000000000014) => RGBA[252, 252, 252, 98]
rgba(252, 252, 252, 0.39990000000000014) => RGBA[252, 252, 252, 102]
rgba(252, 252, 252, 0.41323000000000015) => RGBA[252, 252, 252, 105]
rgba(252, 252, 252, 0.42656000000000016) => RGBA[252, 252, 252, 109]
rgba(252, 252, 252, 0.43989000000000017) => RGBA[252, 252, 252, 112]
rgba(252, 252, 252, 0.4532200000000002) => RGBA[252, 252, 252, 116]
rgba(252, 252, 252, 0.4665500000000002) => RGBA[252, 252, 252, 119]
rgba(252, 252, 252, 0.4798800000000002) => RGBA[252, 252, 252, 122]
rgba(252, 252, 252, 0.4932100000000002) => RGBA[252, 252, 252, 126]
rgba(252, 252, 252, 0.5065400000000002) => RGBA[251, 251, 251, 129]
rgba(252, 252, 252, 0.5198700000000002) => RGBA[251, 251, 251, 133]
rgba(252, 252, 252, 0.5332000000000001) => RGBA[251, 251, 251, 136]
rgba(252, 252, 252, 0.5465300000000001) => RGBA[251, 251, 251, 139]
rgba(252, 252, 252, 0.55986) => RGBA[251, 251, 251, 143]
rgba(252, 252, 252, 0.57319) => RGBA[251, 251, 251, 146]
rgba(252, 252, 252, 0.5865199999999999) => RGBA[251, 251, 251, 150]
rgba(252, 252, 252, 0.5998499999999999) => RGBA[251, 251, 251, 153]
rgba(252, 252, 252, 0.6131799999999998) => RGBA[251, 251, 251, 156]
rgba(252, 252, 252, 0.6265099999999998) => RGBA[251, 251, 251, 160]
rgba(252, 252, 252, 0.6398399999999997) => RGBA[251, 251, 251, 163]
rgba(252, 252, 252, 0.6531699999999997) => RGBA[251, 251, 251, 167]
rgba(252, 252, 252, 0.6664999999999996) => RGBA[252, 252, 252, 170]
rgba(252, 252, 252, 0.6798299999999996) => RGBA[252, 252, 252, 174]
rgba(252, 252, 252, 0.6931599999999996) => RGBA[252, 252, 252, 177]
rgba(252, 252, 252, 0.7064899999999995) => RGBA[252, 252, 252, 180]
rgba(252, 252, 252, 0.7198199999999995) => RGBA[252, 252, 252, 184]
rgba(252, 252, 252, 0.7331499999999994) => RGBA[252, 252, 252, 187]
rgba(252, 252, 252, 0.7464799999999994) => RGBA[252, 252, 252, 191]
rgba(252, 252, 252, 0.7598099999999993) => RGBA[252, 252, 252, 194]
rgba(252, 252, 252, 0.7731399999999993) => RGBA[252, 252, 252, 197]
rgba(252, 252, 252, 0.7864699999999992) => RGBA[252, 252, 252, 201]
rgba(252, 252, 252, 0.7997999999999992) => RGBA[252, 252, 252, 204]
rgba(252, 252, 252, 0.8131299999999991) => RGBA[252, 252, 252, 208]
rgba(252, 252, 252, 0.8264599999999991) => RGBA[252, 252, 252, 211]
rgba(252, 252, 252, 0.839789999999999) => RGBA[251, 251, 251, 214]
rgba(252, 252, 252, 0.853119999999999) => RGBA[251, 251, 251, 218]
rgba(252, 252, 252, 0.8664499999999989) => RGBA[251, 251, 251, 221]
rgba(252, 252, 252, 0.8797799999999989) => RGBA[251, 251, 251, 225]
rgba(252, 252, 252, 0.8931099999999988) => RGBA[251, 251, 251, 228]
rgba(252, 252, 252, 0.9064399999999988) => RGBA[251, 251, 251, 232]
rgba(252, 252, 252, 0.9197699999999988) => RGBA[251, 251, 251, 235]
rgba(252, 252, 252, 0.9330999999999987) => RGBA[251, 251, 251, 238]
rgba(252, 252, 252, 0.9464299999999987) => RGBA[251, 251, 251, 242]
rgba(252, 252, 252, 0.9597599999999986) => RGBA[251, 251, 251, 245]
rgba(252, 252, 252, 0.9730899999999986) => RGBA[251, 251, 251, 249]
rgba(252, 252, 252, 0.9864199999999985) => RGBA[251, 251, 251, 252]
rgba(252, 252, 252, 0.9997499999999985) => RGBA[252, 252, 252, 255]
PASS successfullyParsed is true

TEST COMPLETE
